backend-api/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
pymongo==4.5.0
redis==5.0.1
pika==1.3.2
python-dotenv==1.0.0
